(PC-20415) refactor(fetchVenues): move fetchVenues + test to its own folder
